refactor(suite): split metadata labeling actions
