+++ suite/metadata/src/metadataMiddleware.ts
@@ -4,7 +4,9 @@ import { type AnyAction, createMiddleware } from '@suite-common/redux-utils';
 import { accountsActions, applyDeviceStatesThunk } from '@suite-common/wallet-core';
 
 import { type MetadataAction } from './metadataActions';
-import * as metadataLabelingActions from './metadataLabelingActions';
+import { setAccountMetadataKey } from './metadataLabelingActions/setAccountMetadataKey';
+import { setEditing } from './metadataLabelingActions/setEditing';
+import { selectLabelingValueBeingEdited } from './metadataReducer';
 import * as metadataThunks from './metadataThunks';
 import { selectIsLegacyLabelingVisible } from './selectIsLegacyLabelingVisible';
 
@@ -13,9 +15,7 @@ type Action = MetadataAction | ReturnType<typeof accountsActions.createAccount> 
 export const metadataMiddleware = createMiddleware<Action>(
     (action, { dispatch, getState, next }) => {
         if (accountsActions.createAccount.match(action)) {
-            action.payload = dispatch(
-                metadataLabelingActions.setAccountMetadataKey(action.payload),
-            );
+            action.payload = dispatch(setAccountMetadataKey(action.payload));
         }
 
         if (
@@ -35,8 +35,8 @@ export const metadataMiddleware = createMiddleware<Action>(
         switch (action.type) {
             case '@router/location-change': // hack: to prevent dependency
                 // if there is editing field active, changing route turns it inactive
-                if (getState().metadata.editing) {
-                    dispatch(metadataLabelingActions.setEditing(undefined));
+                if (selectLabelingValueBeingEdited(getState())) {
+                    dispatch(setEditing(undefined));
                 }
                 break;
             default:
